[TIMOB-23911] Remove special property handling
We introduced a workaround in Hyperloop 1.2.7 to remain backwards
compatible with certain properties when transitioning from iOS 9 to
iOS 10. This allowed users to keep using methods instead of  having
to refactor their code to property usage. With Hyperloop 2.0.0 we
introduce this as a breaking change and stick to the new favored way
of using properties, e.g. UIColor.redColor instead of
UIColor.redColor().