[TIMOB-24383] Make sure to use correct prototype for existing wrappers
A wrapper object is stored with the prototype it was first created with. Under certain use cases it can happen that the wrapper for a pointer is generated for a less specific type in the inheritance chain like NSObject. All further access to that wrapper would then use the NSObject prototype even if casted to their specific type, rendering all methods and properties from the specific wrapper implementation unusable. This commit fixes that by making sure to always set the correct prototype before returning an existing wrapper instance.