Hardcode runpath search path (2_2_X) (#224)
* Hardcode runpath search path

Building on Jenkins with an env var results in an invalid runpath search path when running the metabase binary on user machines. Hardcode this for now until we find a solution in TIMOB-23956.

* Check availability of Xcode under default location

The metabase parser binary requires Xcode to be located under /Applications/Xcode.app. This check will enforce the correct Xcode location before Hyperloop can be used to avoid unexpected behavior.