[TIMOB-25401] Improve framework detection (#239)
Fixes a regression that was introduced by changes made to the metabase parser. The metabase parser only detects framework names if a header file comes directly from a .framework package. For all other files, the framewrok will be left as the header path the symbol was found in. This code will iterate over the generated metabase and maps those paths to the appropriate framework name.