[TIMOB-25059] Use module hooks instead of plugins
Moves all the platform specific Hyperloop plugins to the module's hooks folder so we don't need to explicitly register a plugin anymore. Also updates the local build script as well as the Jenkins build to handle the new directory structure. This also gets rid of all the manual plugin packaging.