[TIMOB-25059] Use module hooks instead of plugins (#254)
* [TIMOB-25059] Use module hooks instead of plugins

Moves all the platform specific Hyperloop plugins to the module's hooks folder so we don't need to explicitly register a plugin anymore. Also updates the local build script as well as the Jenkins build to handle the new directory structure. This also gets rid of all the manual plugin packaging.

* [TIMOB-25059] Remove plugin handling from Windows build

* [TIMOB-25059] Build against SDK 7.0.0

* [TIMOB-25059] Remove metabase binary

This shouldn't have been commited in the first place, but the .gitignore wasn't properly updated, duh!

* [TIMOB-25059] Warn about legacy plugin usage

* [TIMOB-25059] Add JS docs

* [TIMOB-25059] Rephrase error message