[TIMOB-25604] Improve compatibility with CocoaPods (#266)
If a Pod is only available as a dynamic framework but the use_frameworks! flag is missing, CocoaPods will create a dummy directory in its public header path and link all headers to the backing framework. This fix prevents duplicate parsing of those headers, which caused some follow up issues within the metabase parser.