fix: exclude arm archs for simulator builds
Building the Hyperloop example app with Xcode 12 GM fails due to linker issues. Building with the beta version seems to work fine. The only way i could find to get it to work with Xcode 12 GM is to add the lines from this PR to the module.xcconfig. They basically exclude arm architectures for simulator builds.